+++ integration_test/pg/test_helper.exs
@@ -103,10 +103,10 @@ defmodule Ecto.Integration.Postgres.Case do
   end
 
   setup do
-    :ok = Ecto.Adapters.Postgres.begin_test_transaction(TestRepo, [])
+    :ok = Ecto.Adapters.SQL.begin_test_transaction(TestRepo, [])
 
     on_exit fn ->
-      :ok = Ecto.Adapters.Postgres.rollback_test_transaction(TestRepo, [])
+      :ok = Ecto.Adapters.SQL.rollback_test_transaction(TestRepo, [])
     end
 
     :ok