+++ test/ecto/adapter/sql_test.exs
@@ -1,3 +1,0 @@
-defmodule Ecto.Adapter.SQLTest do
-  use ExUnit.Case, async: true
-end